TileComponent.vue: renamed.
