+++ src/components/tiles/TileWait.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script lang="ts">
-import Tile from "@/components/tiles/Tile.vue";
+import Tile from "@/components/tiles/TileComponent.vue";
 import { defineComponent } from "vue";
 export default defineComponent({
   components: { Tile },
