+++ src/components/tiles/Tile_PrinterAccount.vue
@@ -15,7 +15,7 @@
 import { defineComponent, onMounted, ref, watchEffect } from "vue";
 import get_content_via_post from "@/helpers/post";
 import { useStore } from "vuex";
-import Tile from "@/components/tiles/Tile.vue";
+import Tile from "@/components/tiles/TileComponent.vue";
 
 export default defineComponent({
   name: "TilePrinterAccount",
