+++ src/views/View_About.vue
@@ -30,7 +30,7 @@
 </template>
 
 <script lang="ts">
-import Tile from "@/components/tiles/Tile.vue";
+import Tile from "@/components/tiles/TileComponent.vue";
 import { defineComponent } from "vue";
 
 const contributors = require("@/assets/contributors.json");
