+++ src/views/View_CampusPlan.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script lang="ts">
-import Tile from "@/components/tiles/Tile.vue";
+import Tile from "@/components/tiles/TileComponent.vue";
 import { defineComponent } from "vue";
 
 export default defineComponent({
