+++ src/views/api_relevant/general/View_Endlicht.vue
@@ -23,7 +23,7 @@
 <script lang="ts">
 import { get } from "@/helpers/fetchData";
 import TileGroup from "@/components/tiles/TileGroup.vue";
-import Tile from "@/components/tiles/Tile.vue";
+import Tile from "@/components/tiles/TileComponent.vue";
 import { defineComponent, onMounted, Ref, ref } from "vue";
 import { weekdays } from "@/helpers/dateHelper";
