+++ src/views/api_relevant/user_related/View_Timetable.vue
@@ -12,7 +12,7 @@
 import { defineComponent, onMounted, ref, watchEffect } from "vue";
 import get_content_via_post from "@/helpers/post";
 import { useStore } from "vuex";
-import Tile from "@/components/tiles/Tile.vue";
+import Tile from "@/components/tiles/TileComponent.vue";
 
 export default defineComponent({
   name: "View_Timetable",
